# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MovieFlix API",
        "description": "API de filmes",
        "version": "1.0.0"
    },
    "paths":{
        "/movies":{
            "get":{
                "tags":["Movie"],
                "summary": "obter todos os filmes do banco de dados",
                "responses":{
                    "200:":{
                        "descriptio": "retorna uma lista de filmes",
                        "content": {
                            "application/json":{
                                "schema":{
                                    "type:": "array",
                                    "items":{
                                        "$ref": "#/components/schemas/Movie"
                                    }
                                }
                            }
                            
                        }
                    }
                }
            }
        },
        "/movies/{id}":{
            "put":{
                "tags": ["Movie"],
                "summary": "Atualizar um filme",
                "parameters":[
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID do filme a ser atualizado",
                        "schema":{
                            "type": "integer"
                        }
                    }
                ],
                "requestBody":{
                    "required": true,
                    "content":{
                        "application/json":{
                            "schema":{
                                "$ref": "#/components/schemas/MovieInput"
                            }
                        }
                    }
                },
                "responses":{
                    "200": {
                        "description": "filme atualizado com sucesso!"
                    },
                    "404": {
                        "description": "Filme não encontrado!"
                    },
                    "500":{
                        "description": "Falha ao atualizar o registro do filme"
                    }
                }
            }
        }
    },
    "components":{
        "schemas":{
            "Movie":{
                "type": "object",
                "properties":{
                    "id": {
                        "type": "integer",
                        "description": "ID do filme no banco de dados"
                    }

                }
            },
            "MovieInput":{
                "type": "object",
                "required": [
                    "title",
                    "genre_id",
                    "language_id",
                    "oscar_count",
                    "release_date"
                ],
                "properties":{
                    "title":{
                        "type": "string",
                        "description": "Nome do título do filme"
                    },
                    "genre_id":{
                        "type": "integer",
                        "description": "ID do gênero do filme"
                    },
                    "language_id": {
                        "type":"integer",
                        "description": "ID da lingua do filme"
                    },
                    "oscar_count":{
                        "type": "integer",
                        "description": "Numero de Oscar's que o filme possue"
                    },
                    "release_date":{
                        "type":"string",
                        "format": "date",
                        "description": "Data de lançamento do filme"
                    }
                }
            }
        }
    }
    
}